refactor(article,series): integrate series rdb adapter to article rdb adapter
