+++ services/article/driven/rdb/src/main/java/nettee/blolet/article/rdb/mapper/ArticleLikesEntityMapper.java
@@ -1,0 +1,11 @@
+package nettee.blolet.article.rdb.mapper;
+
+import nettee.blolet.article.domain.ArticleLikes;
+import nettee.blolet.article.rdb.entity.ArticleLikesEntity;
+import org.mapstruct.Mapper;
+
+@Mapper(componentModel = "spring")
+public interface ArticleLikesEntityMapper {
+    ArticleLikesEntity toEntity(ArticleLikes domain);
+    ArticleLikes toDomain(ArticleLikesEntity entity);
+}
