+++ services/article/driven/rdb/src/test/java/nettee/blolet/article/rdb/SeriesRdbTestApplication.java
@@ -1,0 +1,11 @@
+package nettee.blolet.article.rdb;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class SeriesRdbTestApplication {
+    public static void main(String[] args) {
+        SpringApplication.run(SeriesRdbTestApplication.class, args);
+    }
+}
