--- services/series/driven/rdb/src/main/java/nettee/series/article/driven/rdb/persistence/SeriesArticleJpaRepository.java
@@ -1,18 +1,0 @@
-package nettee.series.article.driven.rdb.persistence;
-
-import nettee.series.article.driven.rdb.entity.SeriesArticleEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Modifying;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
-import org.springframework.transaction.annotation.Transactional;
-
-public interface SeriesArticleJpaRepository extends JpaRepository<SeriesArticleEntity, Long> {
-    
-    SeriesArticleEntity findBySeriesIdAndDraftId(Long seriesId, Long draftId);
-    
-    @Modifying
-    @Transactional
-    @Query("delete from SeriesArticleEntity s where s.seriesId = :seriesId")
-    void deleteAllBySeriesId(@Param("seriesId") Long seriesId);
-}
