--- services/series/driven/rdb/src/main/java/nettee/series/driven/rdb/persistence/SeriesJpaRepository.java
@@ -1,9 +1,0 @@
-package nettee.series.driven.rdb.persistence;
-
-import nettee.series.driven.rdb.entity.SeriesEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface SeriesJpaRepository extends JpaRepository<SeriesEntity, Long> {
-    
-    boolean existsByBlogIdAndTitle(Long blogId, String title);
-}
