--- services/series/driven/rdb/src/test/java/nettee/series/SeriesRdbTestApplication.java
@@ -1,11 +1,0 @@
-package nettee.series;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class SeriesRdbTestApplication {
-    public static void main(String[] args) {
-        SpringApplication.run(SeriesRdbTestApplication.class, args);
-    }
-}
