--- services/series/driven/rdb/src/test/kotlin/nettee/series/driven/rdb/jpa/JpaTransactionalFreeSpec.kt
@@ -1,13 +1,0 @@
-package nettee.series.driven.rdb.jpa
-
-import io.kotest.core.spec.style.FreeSpec
-import io.kotest.extensions.spring.SpringExtension
-import nettee.jpa.config.JpaConfig
-import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase
-import org.springframework.context.annotation.ComponentScan
-
-@AutoConfigureTestDatabase(replace = AutoConfigureTestDatabase.Replace.NONE)
-@ComponentScan(basePackageClasses = [JpaConfig::class])
-abstract class JpaTransactionalFreeSpec(body: FreeSpec.() -> Unit) : FreeSpec(body) /* super(body) */ {
-    override fun extensions() = listOf(SpringExtension)
-}
